+++ lib/minitest-activemodel/matchers/activemodel/validate_length_matcher.rb
@@ -4,10 +4,12 @@ module MiniTest
       # Ensures that the length/size of the attribute is validated.
       #
       # Options:
-      # * <tt>with_minimum</tt> - minimum length of the attribute.
+      # * <tt>with_minimum</tt> - the minimum length of the attribute.
       #   Aliased as: <tt>with_min</tt> and <tt>is_at_least</tt>.
-      # * <tt>with_maximum</tt> - maximum length of the attribute.
+      # * <tt>with_maximum</tt> - the maximum length of the attribute.
       #   Aliased as: <tt>with_max</tt> and <tt>is_at_most</tt>.
+      # * <tt>within</tt> - the rage specifying the minimum and maximum length
+      #   of the attribute.
       #
       #   it { must validate_length_of :name }
       #   it { must validate_length_of(:name).with_minimum(10) }
@@ -16,6 +18,7 @@ module MiniTest
       #   it { must validate_length_of(:name).with_maximum(100) }
       #   it { must validate_length_of(:name).with_max(100) }
       #   it { must validate_length_of(:name).is_at_most(100) }
+      #   it { must validate_length_of(:name).within(10..100) }
       def validate_length_of attr
         ValidateLengthMatcher.new attr
       end
@@ -44,19 +47,26 @@ module MiniTest
         alias :with_max   :with_maximum
         alias :is_at_most :with_maximum
 
+        def within value
+          @within = value
+          self
+        end
+
         def matches? subject
           return false unless @result = super(subject)
 
           check_minimum if @minimum
           check_maximum if @maximum
+          check_range   if @within
 
           @result
         end
 
         def description
           desc = []
-          desc << "with minimum #{@minimum}" if @minimum
+          desc << " with minimum #{@minimum}" if @minimum
           desc << " with maximum #{@maximum}" if @maximum
+          desc << " within range #{@within}"  if @within
 
           super << desc.to_sentence
         end
@@ -81,6 +91,19 @@ module MiniTest
             @positive_message << " with maximum of #{actual}"
           else
             @negative_message << " with maximum of #{actual}"
+            @result = false
+          end
+        end
+
+        def check_range
+          min, max = @within.min, @within.max
+          actual_min = @validator.options[:minimum]
+          actual_max = @validator.options[:maximum]
+
+          if actual_min == min && actual_max == max
+            @positive_message << " with range #{@within}"
+          else
+            @negative_message << " with range #{actual_min..actual_max}"
             @result = false
           end
         end