+++ CHANGELOG.md
@@ -11,3 +11,4 @@
 + Add `with_min` and `is_at_least` aliases to `ValidateLengthMatcher#with_minimum`.
 + Add `with_maximum` option to `ValidateLengthMatcher`.
 + Add `with_max` and `is_at_most` aliases to `ValidateLengthMatcher#with_maximum`.
++ Add `within` option to `ValidateLengthMatcher`.